Refactor import statement for AssetHandler
