+++ chat_exporter/__init__.py
@@ -1,5 +1,4 @@
 from chat_exporter.chat_exporter import export, raw_export, quick_export, link, quick_link
-from chat_exporter.construct.asset_handler import AssetHandler
 
 __version__ = "2.7.0"
